refacto: rename Answer component
