+++ ui/components/textCanvas/src/components/SpannableMessage.svelte
