+++ ui/components/textCanvas/src/components/index.ts
@@ -4,4 +4,4 @@ Author : pixano@cea.fr
 License: CECILL-C
 -------------------------------------*/
 
-export { default as Answer } from "./Answer.svelte";
+export { default as SpannableMessage } from "./SpannableMessage.svelte";
